refactor: Used async for all data get functions, split bloons.py code
