+++ bot/cogs/TilestratCog.py
@@ -8,7 +8,16 @@ import bot.utils.discordutils
 from bot.exceptions import UnknownTile
 from bot.classes import ErrorHandlerCog
 from bot.views import EmbedPaginateView
-import bot.utils.bloons
+from bot.utils.bloons import (
+    raw_challenge_to_embed,
+    get_current_ct_number,
+)
+from bot.utils.misc import add_spaces
+from bot.utils.bloonsdata import (
+    fetch_tile_data,
+    relic_to_tile_code,
+    fetch_all_tiles,
+)
 from bot.utils.emojis import LEAST_TIERS, LEAST_CASH, BLOONARIUS, VORTEX, LYCH, TIME_ATTACK, BLANK, DREADBLOON, \
     PHAYZE
 from bot.utils.images import IMG_LEAST_CASH, IMG_LEAST_TIERS, IMG_BLOONARIUS, IMG_VORTEX, IMG_LYCH, IMG_TIME_ATTACK, \
@@ -138,7 +147,7 @@ class TilestratCog(ErrorHandlerCog):
         tile_code = tile_code.strip().upper()
         tile_re = r"(?:M|[A-G])(?:R|[A-G])(?:X|[A-H])"
         if len(tile_code) != 3 or re.match(tile_re, tile_code) is None:
-            actual_tile_code = await asyncio.to_thread(bot.utils.bloons.relic_to_tile_code, tile_code)
+            actual_tile_code = await relic_to_tile_code(tile_code)
             if actual_tile_code is None:
                 raise UnknownTile(tile_code)
             tile_code = actual_tile_code
@@ -149,7 +158,7 @@ class TilestratCog(ErrorHandlerCog):
             await bot.db.queries.tilestrat.del_tile_strat_forum(interaction.guild_id)
             raise TilestratForumNotFound()
 
-        tile_info = await asyncio.to_thread(bot.utils.bloons.fetch_tile_data, tile_code)
+        tile_info = await fetch_tile_data(tile_code)
         if tile_info is None:
             raise UnknownTile(tile_code)
 
@@ -200,7 +209,7 @@ class TilestratCog(ErrorHandlerCog):
                 boss_name_tag = await forum_channel.create_tag(name=boss_name)
             tags.append(boss_name_tag)
 
-        map_name = bot.utils.bloons.add_spaces(tile_data["selectedMap"])
+        map_name = add_spaces(tile_data["selectedMap"])
         if map_name == "Adoras Temple":
             map_name = "Adora's Temple"
         elif map_name == "Pats Pond":
@@ -213,7 +222,7 @@ class TilestratCog(ErrorHandlerCog):
             name=thread_template.format(event_num=tile_info["EventNumber"], map=map_name, tile_code=tile_info["Code"]),
             content=thread_init_message,
             applied_tags=tags,
-            embed=bot.utils.bloons.raw_challenge_to_embed(tile_info),
+            embed=raw_challenge_to_embed(tile_info),
         )).thread
         await self.on_raidlog_created(thread, tile_info, forum_channel.id)
         return thread
@@ -233,7 +242,7 @@ class TilestratCog(ErrorHandlerCog):
             raise TilestratForumNotFound()
 
         if season is None:
-            season = bot.utils.bloons.get_current_ct_number()
+            season = get_current_ct_number()
 
         logged_count = {
             "Banner": {"race": 0, "lc": 0, "lt": 0, "boss": 0},
@@ -246,7 +255,7 @@ class TilestratCog(ErrorHandlerCog):
             "Regular": {"race": 0, "lc": 0, "lt": 0},
         }
 
-        tiles = await asyncio.to_thread(bot.utils.bloons.fetch_all_tiles)
+        tiles = await fetch_all_tiles()
         logged_tiles = await bot.db.queries.tilestrat.get_tilestrats_by_season(season, forum_id)
         logged_tiles = [lt.tile_code for lt in logged_tiles]
