+++ config.example.py
@@ -6,6 +6,8 @@ DB_PSWD = "postgres"
 DB_HOST = "127.0.0.1"
 DB_NAME = "ct_ticket_tracker"
 
+DATA_PATH = "~"
+
 # Guild with emote perms for the bot and all 100 emote slots free. Needed for the leaderboard cog.
 # Leave blank if you don't care about the leaderboard having team icon emotes.
 EMOTE_GUILD_ID = None
