+++ pandehelper.py
@@ -3,7 +3,7 @@ import discord
 import logging
 from datetime import datetime
 import bot.db.connection
-import bot.utils.bloons
+import bot.utils.bloonsdata
 from bot import __version__
 from discord.ext import commands
 from config import TOKEN, APP_ID
@@ -29,7 +29,7 @@ class CtTicketTracker(commands.Bot):
             os.mkdir("tmp")
 
     async def setup_hook(self):
-        await bot.utils.bloons.init_bloonspy_client()
+        await bot.utils.bloonsdata.init_bloonspy_client()
         await bot.db.connection.start()
         cogs = [
             "OwnerCog",
